feat: extend client runtime for headless world replication
- Keep the client replication system active in headless worlds without creating client physics or presentation state.
- Add headless snapshot handling with baseline tracking and full-baseline recovery for missing or mismatched deltas.
- Skip lifecycle and ECS actor processing for headless clients while preserving replication acknowledgements.
- Allocate process-wide unique pipeline subtypes for concurrently active client worlds.
- Preserve character control revision ordering across queued input commands.
- Expose the client instance ID through ClientRuntime for multi-client hosting.
- Detect actor-directory generation conflicts and safely reclaim stale out-of-AOI occupants.
- Improve lifecycle packet validation and diagnostic warnings for malformed or unavailable replication paths

+++ JamNet/include/jamnet/runtime/world/simulation/client/ClientWorld.h
@@ -37,7 +37,7 @@ namespace jam::net
 		void								Stop() override;
 		void								SetHeadless(bool headless) { m_headless = headless; }
 		bool								IsHeadless() const { return m_headless; }
-		void								SetPipelineSubtype(uint16 subtype) { m_pipelineSubtype = subtype; }
+		void								SetPipelineSubtype(uint32 subtype) { m_pipelineSubtype = subtype; }
 
 		void								SetPrincipalState(ClientPrincipalState* principal) { m_principal = principal; }
 		uint64								GetAccountId() const { return m_principal ? m_principal->accountId : kInvalidAccountId; }
@@ -96,9 +96,8 @@ namespace jam::net
 	private:
 		ClientPrincipalState*				m_principal		  = nullptr;
 
-		uint16								m_pipelineSubtype = 0;
+		uint32								m_pipelineSubtype = 0;
 		bool								m_headless = false;
 		ActorId								m_localActorId = ActorId::Invalid();
-
 	};
 }
